fix: resolve pipe not rendering when changing graph
