+++ src/renderer/components/AsyncRenderer.tsx
@@ -16,7 +16,8 @@ const AsyncComponent: FC<{
     renderer: () => Promise<any> | any;
     loadingContent: ReactElement | string;
     delay?: number;
-}> = ({ renderer, loadingContent: fallback, delay = 0 }) => {
+    postRenderCallback?: () => any;
+}> = ({ renderer, loadingContent: fallback, delay = 0, postRenderCallback }) => {
     const [content, setContent] = useState(fallback);
 
     // This effect is so we can have both micro and micro tasks queued.
@@ -25,6 +26,7 @@ const AsyncComponent: FC<{
     useEffect(() => {
         (async () => {
             setContent(await asyncRenderer(renderer, delay));
+            postRenderCallback?.();
         })();
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [renderer]);
